+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 
 import './globals.css'
-import { AuthProvider } from '@/components/auth/AuthProvider'
+import AuthProvider from '@/components/auth/AuthProvider'
 import { pretendard } from '@/styles/fonts'
 
 export const metadata: Metadata = {
@@ -34,11 +34,7 @@ export const metadata: Metadata = {
   },
 }
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+const RootLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html lang="ko" className={pretendard.variable}>
       <head>
@@ -57,3 +53,5 @@ export default function RootLayout({
     </html>
   )
 }
+
+export default RootLayout
