+++ src/providers/ReactQueryProvider.tsx
@@ -3,11 +3,11 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { useState } from 'react'
 
-export default function ReactQueryProvider({
+export const ReactQueryProvider = ({
   children,
 }: {
   children: React.ReactNode
-}) {
+}) => {
   const [queryClient] = useState(
     () =>
       new QueryClient({
